Conflicts:
	Server/Client.cpp
	Server/Client.h

+++ Server/Wall.cpp
@@ -5,7 +5,7 @@ const float Wall::sphereRadius = 5.0f;
 Wall::Wall() {
   //TODO initialize
 }
-Wall::Wall(unsigned int w, unsigned int d, unsigned int h, v3_t sc, v3_t direct, Map * m):centerPositions()
+Wall::Wall(float w, float d, float h, v3_t sc, v3_t direct, Map * m):centerPositions()
 {
   width = w;
   depth = d;
